Fix autocomplete by id. Fixes #380.
